So much refactoring that I got dizzy, plus some of the config stuff
